Make SonataFlow images deps available offline, fix jbang setup and deps
