+++ tests/shell/kogito-swf-builder/.vscode/settings.json
@@ -1,0 +1,18 @@
+{
+    "java.import.gradle.enabled": false,
+    "java.import.maven.enabled": false,
+    "java.eclipse.downloadSources": true,
+    "files.exclude": {
+        "bin/": true,
+        ".eclipse/": true,
+        ".project": true,
+        ".classpath": true,
+        "build.gradle": true
+    },
+    "java.completion.importOrder": [
+        "",
+        "javax",
+        "java",
+        "#"
+    ]
+}
