+++ tests/shell/kogito-swf-builder/resources/greet/Dockerfile
@@ -19,7 +19,7 @@ RUN "${KOGITO_HOME}"/launch/build-app.sh './resources'
 #=============================
 # Runtime Run
 #=============================
-FROM registry.access.redhat.com/ubi8/openjdk-11-runtime:latest
+FROM registry.access.redhat.com/ubi8/openjdk-17-runtime:latest
 
 ARG QUARKUS_LAUNCH_DEVMODE=false
