+++ tests/shell/kogito-swf-devmode/.vscode/launch.json
@@ -1,0 +1,22 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "type": "java",
+            "name": "RunTests-port-4004",
+            "request": "attach",
+            "hostName": "localhost",
+            "port": 4004
+        },
+        {
+            "type": "java",
+            "name": "Debug (Launch) - RunTests",
+            "request": "launch",
+            "mainClass": "RunTests",
+            "projectName": "RunTests"
+        }
+    ]
+}
