+++ tests/shell/kogito-swf-devmode/src/RunTests.java
@@ -1,21 +1,20 @@
 ///usr/bin/env jbang "$0" "$@" ; exit $?
 
-//DEPS org.slf4j:slf4j-simple:2.0.6
+//DEPS org.slf4j:slf4j-simple:2.0.9
 
 // Junit console to start the test engine:
-//DEPS org.junit.platform:junit-platform-console:1.8.2
+//DEPS org.junit.platform:junit-platform-console:1.10.1
 
 // engine to run the tests (tests are written with Junit5):
-//DEPS org.junit.jupiter:junit-jupiter-engine:5.8.2
+//DEPS org.junit.jupiter:junit-jupiter-engine:5.10.1
 
 // testcontainers
-//DEPS org.testcontainers:testcontainers:1.17.6
-//DEPS org.testcontainers:junit-jupiter:1.17.6
+//DEPS org.testcontainers:testcontainers:1.19.3
+//DEPS org.testcontainers:junit-jupiter:1.19.3
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.http.HttpClient;
@@ -24,18 +23,14 @@ import java.net.http.HttpResponse;
 import java.net.http.HttpResponse.BodyHandlers;
 import java.nio.file.Paths;
 import java.time.Duration;
-import java.util.Collections;
-
 import org.junit.jupiter.api.Test;
-import org.junit.platform.console.options.CommandLineOptions;
-import org.junit.platform.console.tasks.ConsoleTestExecutor;
+import org.junit.platform.console.ConsoleLauncher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testcontainers.containers.BindMode;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.output.Slf4jLogConsumer;
 import org.testcontainers.containers.wait.strategy.Wait;
-import org.testcontainers.images.builder.ImageFromDockerfile;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 
@@ -48,16 +43,20 @@ public class RunTests {
 
     @Container
     private GenericContainer devModeImage = new GenericContainer(getTestImage())
-            .withFileSystemBind(getScriptDirPath() + "/resources", "/home/kogito/serverless-workflow-project/src/main/resources", BindMode.READ_ONLY)
+            .withEnv("MAVEN_ARGS_APPEND", "-Ddebug=false -Dquarkus.devservices.enabled=false")
+            .withFileSystemBind(getScriptDirPath() + "/resources",
+                    "/home/kogito/serverless-workflow-project/src/main/resources", BindMode.READ_ONLY)
             .withExposedPorts(8080)
             .waitingFor(Wait.forHttp("/jsongreet"))
+            .withStartupTimeout(Duration.ofMinutes(2))
             .withLogConsumer(logConsumer);
 
     @Test
     public void testBuiltContainerAnswerCorrectly() throws URISyntaxException, IOException, InterruptedException {
         devModeImage.start();
         HttpRequest request = HttpRequest.newBuilder()
-                .uri(new URI("http://" + devModeImage.getHost() + ":" + devModeImage.getFirstMappedPort() + "/jsongreet"))
+                .uri(new URI(
+                        "http://" + devModeImage.getHost() + ":" + devModeImage.getFirstMappedPort() + "/jsongreet"))
                 .header("Content-Type", "application/json")
                 .header("Accept", "application/json")
                 .timeout(Duration.ofSeconds(10))
@@ -69,10 +68,8 @@ public class RunTests {
     }
 
     public static void main(String... args) throws Exception {
-        CommandLineOptions options = new CommandLineOptions();
-        options.setSelectedClasses(Collections.singletonList(RunTests.class.getName()));
-        options.setReportsDir(Paths.get(getOutputDir()));
-        new ConsoleTestExecutor(options).execute(new PrintWriter(System.out));
+        ConsoleLauncher.main("--select-class=" + RunTests.class.getName(),
+                "--reports-dir=" + Paths.get(getOutputDir()).toString());
     }
 
     static String getTestImage() {
